Extracted StatusFlatList
